Create Assignment_2_V1.0
Creating now Assignment. Defining functions for all of the tasks 1-5. Attempted to implement monitering system given on Canvas but there are errors when I attempt to verify so this must not be implemented correctly. There is also an error in the frequnecy display in task 5. Delays still need to be calculated for the main loop in order to run the executive properly.