Update Assignment_2_V1.0
Defined a cyclic executive loop in terms of a 4ms frame. If statement defines which frames perform which task. A further if else statment will need to be added which performs a delay equal to the amount of time taken to perform the task when the task is not being performed. Also fixed the analog read pins and commands for task 4 which now works correctly. Testing still to be done on tasks 1,2 and 3. Task 5 displays values in the serial moniter. There is the potential to remove extra decimals points where they are not needed.