Update Assignment_2_V1.0
Update made to the cyclic executive. All the tasks now function as as they should and have been measured. There still needs to be a function put in in order to calculate the delay for tasks 2 and 3 which will change with a change in frequency. There is potential to still add a ticker and use case statements as in the example.