Update Assignment_2_V1.0
Using for loops does not work with the ticker becausue it would have to run at total of 50 times each taking 4us so the ticker would only be usefull in ticks of 200us. Using case statements instead works with the ticker. In order to get the Task1() not to miss its first mark you run it once before the start.monitoring and then wait a fixed delay so that it does not overlap with the loop. Also added a cosntrain function in task5 to constrain the output values to 0 and 99.99.